chore(mcp): move generate-app-url eval to ci/eval_basic.py
Relocate eval_app_link_generation (person/event link generation via generate-app-url) from the cli_mcp workflow-conventions suite to the basic-functionality suite, where it fits better — it's mode-agnostic and not about exec-cli conventions. Scorers stay in cli_mcp.scorers and are imported.
